CoursesMenu class
